+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Page Viewer:
 - Functional rendering on/off toggle
 - When off the UI will now render text background and foreground colour correctly, and will render border decorations
   for photos and texts.
+- Better handling of saved slot cropping/scaling when optimising layout
 
 ## [1.0.0] - 2026-01-20